Add more translations and fix some bugs
The title of the start step is translated. In order for that to work, a
few bugs are fixed:
1) We shall use ugettext_lazy() instead of ugettext().
2) When new call to ugettext_lazy() is added, the command
'python manage.py makemessages -l zh_Hans -a' is executed to collect that
into the django.po file. When translation is done, the lines that start
with '#, fuzzy' and '#| msgid' shall be removed for that translation to
take effect. For more info see the gettext manual on format of PO files:
https://www.gnu.org/software/gettext/manual/html_node/PO-Files.html