Initial REST API for per-app task assignment
The design of per-app task management REST API is that, each of the flow
view mixin's url() method is changed from a property function to a
ordinary function accepting an optional keyword argument rest, which is
a boolean value defaulting to False (it is set to True within
viewflow.flow.viewset_rest.FlowViewSet.urls()). The flow view mixin
classes' url() method will pass the value down to the as_view() function
of the view class (still the 'traditional' one as usual). If the flag is
False, the latter will forward call to its super as usual; otherwise, it
will forward the call to a REST version of the view class.

The REST view class is designed as shadow behind the 'traditional' view
class, such that nodes definition in viewflow.flow.nodes are kept
intact.

In this initial version, only AssignTaskView (task assignment) is
implemented.