REST API: implements UpdateProcessRestView
It is observed that UpdateProcessRestView, UpdateProcessRestView, and
OrderCompleteProjectRestView are all doing largely the same things,
except that the former two work on process class with business logic
model merged, while the latter work on process with external business
logic model.

This commit distills the common part and creates a new
UpdateFieldsRestViewMixin from which all the three views are derived.

Following the idea, StartViewRest is created.

In addition, more applications of APIViewWithoutCSRFEnforcement are
introduced.