REST API: enable all four demo apps
The demo apps are previously disabled to make the development easier.
Now that the core REST API is up, it is time to bring them back. Now for
each of them, 'traditional' and REST API end points co-exist together
seamlessly and trasparently.

We could see how easy it is to enable REST API support for them.

In addition, the core view mixin UpdateFieldsRestViewMixin is enhanced
to
1) tolerate more special conditions.
2) fix the problem that the business object is not saved to the process
via the foreign key.
3) write to nested object via serializer.update().